move api pictures in folder API + call from index
